relocate contents of src/lib to src/utils
